pyproject.toml
[tool.poetry]
name = "abstract-telegram-bot"
version = "0.1.0"
description = ""
authors = ["yogson <yogson@gmail.com>"]
readme = "README.md"
packages = [{include = "abstract_telegram_bot"}]

[tool.poetry.dependencies]
python = "3.11.1"
python-telegram-bot = "20.7"
openai = "1.3.6"
uvloop = "0.19.0"
ruff = "^0.1.6"

[tool.ruff]
line-length = 120


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
